Improve performance by changing is_dir semantics
Previously, files (=non-dirs) were queried for is_dir(...) twice when a
directory was loaded. The reason for this was that MotherFileSystem
could not cache the result when is_dir(...) returned False, because a
False result might also have meant that the file does not exist. Now,
is_dir(...) raises FileNotFoundError when the file does not exist. This
lets MotherFileSystem cache False as well.