Add DirectoryPaneListener#before_location_change
Note: This implementation does not correctly set the sort column when
fman is started. To be fixed in a future commit.

The new method will allow changing location and setting the sort column
as one atomic operation. This removes the flickering when changing to a
location with a non-default sort order.